Remove all traces of [node_valid].
This is an absolutely unnecessary proof obligation, which simply
complicated proofs, and added unnecessary stuff into the context. The
original idea was that if you had a node, then you'd be able to easily
get the containment of the two sides. However, this has turned out to
be more painful than anything else, since we are forced to prove
containment everywhere, and this involves a whole bunch of manual
proofs for containment, which we should've gotten for free.

Currently, this breaks the proof for [tot_dll_to_fragment_split],
which needs to be re-looked at, but otherwise everything else seems to
be perfectly fine.

Actually some of the other proofs might be simplified with these
changes, but that isn't something I shall spend time on now.