No need of giving any [ifuel] for these. SEE BELOW.
The proofs go through just fine with a lower [ifuel]. However, this is
the minimum fuel that they need.

Additionally, in regards to [rlimit]s, the issue is that commits
[662db41a81f551] and [eed8f68581923e7f] in F* updated the patterns in
the [*.Modifies] libraries, and added patterns to [loc_union_comm] and
[loc_union_assoc]. These end up triggering _massively_ (31587 & 818503
extra instantiations, without-hints compared to with-hints,
respectively), in just the [dll_insert_at_head]. [dll_insert_at_tail]
has it even worse. This is what causes the proofs to take ~500 seconds
to run (and still be able to prove in 0.3 seconds with hints).

By removing the respective [SMTPat]s from [LowStar.Modifies], the
proof timings come down to just 10 seconds (without hints).

For now, I am continuing the proofs with the changed file (i.e., with
the [SMTPat]s removed), but discussion on the same is ongoing on
the Everest Slack (channel #fstar-hackery).

Relevant link:
https://everestexpedition.slack.com/archives/C0N8CV6NA/p1528842536000109