Add in modification postconditions (see message)
Some of these are much more generous than they need to be. Instead, it
might be a good idea to limit them by throwing in
conditionals. However, I am not entirely sure how much is necessary
here. Instead, it might just be good to (during the interface exposing
phase) simply have the modification happen "on the list".