Make it easier to be able to make a [node_not_in_dll]. See msg.
Turns out, the previous definition is practically the same as the new
one. Turns out, both are practically identical, and in fact, it should
be possible to write one in terms of the other. One is easier to
prove, and the other is easier to use. Fortunately, all the proofs go
through, even if we just use the shorter one, which means that we
trade-off proving time for easiness of usage of the library functions.

It might be a good idea to write a separate lemma that triggers off
immediately whenever [node_not_in_dll] shows up, that brings back the
previous definition.