Finish proving [dll_remove_mid] iface impl. See msg.
And with that, I am done proving everything in iface-impl. All that
remains for this (insanely) long ordeal to finish is to finish up the
few remaining [assume]s and [admit]s within the impl itself. All of
these are to do with the 2 largest functions in there. In particular,
the functional correctness of these functions needs to be shown.

For both of them, I believe it is necessary to invoke the "uniqueness"
lemma (i.e., there are no duplicates in a nodelist). With that being
invoked, I think it should be relatively straightforward (except for
fighting with F*) to finish off these proofs.

Damn this has taken forever (!)